:recycle: Merge base FlowNodeHandlers First step of the refactoring. The handler is far from pretty now, but at least all base functionality is now in the same place, giving us a detailed view over the tasks that the base handler has to perform.
