+++ src/runtime/flow_node_handler/empty_activity_handler.ts
@@ -15,11 +15,11 @@ import {
 } from '@process-engine/process_engine_contracts';
 import {Model} from '@process-engine/process_model.contracts';
 
-import {FlowNodeHandlerInterruptible} from './index';
+import {FlowNodeHandler} from './index';
 
 // This type of handler doesn't actually do anything but pass on the token it receives.
 // Think of it as a kind of break point.
-export class EmptyActivityHandler extends FlowNodeHandlerInterruptible<Model.Activities.EmptyActivity> {
+export class EmptyActivityHandler extends FlowNodeHandler<Model.Activities.EmptyActivity> {
 
   private emptyActivitySubscription: Subscription;
