+++ src/runtime/flow_node_handler/intermediate_event_handlers/intermediate_message_catch_event_handler.ts
@@ -14,9 +14,9 @@ import {
 } from '@process-engine/process_engine_contracts';
 import {Model} from '@process-engine/process_model.contracts';
 
-import {FlowNodeHandlerInterruptible} from '../index';
+import {FlowNodeHandler} from '../index';
 
-export class IntermediateMessageCatchEventHandler extends FlowNodeHandlerInterruptible<Model.Events.IntermediateCatchEvent> {
+export class IntermediateMessageCatchEventHandler extends FlowNodeHandler<Model.Events.IntermediateCatchEvent> {
 
   private subscription: Subscription;
