+++ src/runtime/flow_node_handler/intermediate_event_handlers/intermediate_timer_catch_event_handler.ts
@@ -13,9 +13,9 @@ import {
 } from '@process-engine/process_engine_contracts';
 import {Model} from '@process-engine/process_model.contracts';
 
-import {FlowNodeHandlerInterruptible} from '../index';
+import {FlowNodeHandler} from '../index';
 
-export class IntermediateTimerCatchEventHandler extends FlowNodeHandlerInterruptible<Model.Events.IntermediateCatchEvent> {
+export class IntermediateTimerCatchEventHandler extends FlowNodeHandler<Model.Events.IntermediateCatchEvent> {
 
   private _timerFacade: ITimerFacade;
   private timerSubscription: Subscription;
