+++ src/runtime/flow_node_handler/service_task_handlers/service_task_factory.ts
@@ -3,7 +3,7 @@ import {IContainer} from 'addict-ioc';
 import {IFlowNodeHandler, IFlowNodeHandlerDedicatedFactory} from '@process-engine/process_engine_contracts';
 import {Model} from '@process-engine/process_model.contracts';
 
-import {FlowNodeHandlerInterruptible} from '../flow_node_handler_interruptible';
+import {FlowNodeHandler} from '../flow_node_handler';
 
 export class ServiceTaskFactory implements IFlowNodeHandlerDedicatedFactory<Model.Activities.ServiceTask> {
 
@@ -16,9 +16,9 @@ export class ServiceTaskFactory implements IFlowNodeHandlerDedicatedFactory<Mode
   public async create(flowNode: Model.Activities.ServiceTask): Promise<IFlowNodeHandler<Model.Activities.ServiceTask>> {
 
     if (flowNode.type === Model.Activities.ServiceTaskType.external) {
-      return this._container.resolveAsync<FlowNodeHandlerInterruptible<Model.Activities.ServiceTask>>('ExternalServiceTaskHandler', [flowNode]);
+      return this._container.resolveAsync<FlowNodeHandler<Model.Activities.ServiceTask>>('ExternalServiceTaskHandler', [flowNode]);
     }
 
-    return this._container.resolveAsync<FlowNodeHandlerInterruptible<Model.Activities.ServiceTask>>('InternalServiceTaskHandler', [flowNode]);
+    return this._container.resolveAsync<FlowNodeHandler<Model.Activities.ServiceTask>>('InternalServiceTaskHandler', [flowNode]);
   }
 }
